requirements.txt
alembic==1.0.1
async-timeout==3.0.1
asyncpg==0.24.0
attrs==21.2.0
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.8
gino==1.0.1
greenlet==1.1.1
idna==3.2
importlib-resources==5.2.2
iniconfig==1.1.1
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
multidict==5.1.0
packaging==21.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.15.1
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.2.16
toml==0.10.2
typing-extensions==3.10.0.2
webargs==8.0.1
yarl==1.6.3
zipp==3.5.0
